--- RichTextControls/Html2XHtml/Html2XHtml.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using HtmlAgilityPack;
-
-namespace Html2XHtml
-{
-    public class Html2XHtml
-    {
-        public static string Convert(string Html)
-        {
-            HtmlDocument htmlDocument = new HtmlDocument();
-            htmlDocument.LoadHtml(Html);
-            return htmlDocument.ParsedText;
-        }
-    }
-}
