--- RichTextControls/Html2XHtml/Html2XHtml.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Reference Include="InkingNewstand.HtmlAgilityPack">
-      <HintPath>..\..\..\InkingNewstand\InkingNewstand.Utilities\dlls\InkingNewstand.HtmlAgilityPack\InkingNewstand.HtmlAgilityPack.dll</HintPath>
-    </Reference>
-  </ItemGroup>
-
-</Project>
