+++ RichTextControls/RichTextControls.ReadersExample/App.xaml
@@ -1,8 +1,7 @@
 ﻿<Application
-    x:Class="RichTextControls.Tests.App"
+    x:Class="RichTextControls.ReadersExample.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:RichTextControls.Tests"
-    RequestedTheme="Light">
+    xmlns:local="using:RichTextControls.ReadersExample">
 
 </Application>
