+++ RichTextControls/RichTextControls.ReadersExample/Package.appxmanifest
@@ -1,19 +1,21 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
+
 <Package
   xmlns="http://schemas.microsoft.com/appx/manifest/foundation/windows10"
   xmlns:mp="http://schemas.microsoft.com/appx/2014/phone/manifest"
   xmlns:uap="http://schemas.microsoft.com/appx/manifest/uap/windows10"
   IgnorableNamespaces="uap mp">
 
-  <Identity Name="e60af4b9-fb31-49ed-9922-c6186201e4a0"
-            Publisher="CN=Ryan"
-            Version="1.0.0.0" />
+  <Identity
+    Name="b385ca7c-cd41-4d65-a464-b0d680f2550b"
+    Publisher="CN=redal"
+    Version="1.0.0.0" />
 
-  <mp:PhoneIdentity PhoneProductId="e60af4b9-fb31-49ed-9922-c6186201e4a0" PhonePublisherId="00000000-0000-0000-0000-000000000000"/>
+  <mp:PhoneIdentity PhoneProductId="b385ca7c-cd41-4d65-a464-b0d680f2550b" PhonePublisherId="00000000-0000-0000-0000-000000000000"/>
 
   <Properties>
-    <DisplayName>RichTextControls.Tests</DisplayName>
-    <PublisherDisplayName>Ryan</PublisherDisplayName>
+    <DisplayName>RichTextControls.ReadersExample</DisplayName>
+    <PublisherDisplayName>redal</PublisherDisplayName>
     <Logo>Assets\StoreLogo.png</Logo>
   </Properties>
 
@@ -22,24 +24,26 @@
   </Dependencies>
 
   <Resources>
-    <Resource Language="x-generate" />
+    <Resource Language="x-generate"/>
   </Resources>
+
   <Applications>
-    <Application Id="vstest.executionengine.universal.App"
-        Executable="$targetnametoken$.exe"
-        EntryPoint="RichTextControls.Tests.App">
+    <Application Id="App"
+      Executable="$targetnametoken$.exe"
+      EntryPoint="RichTextControls.ReadersExample.App">
       <uap:VisualElements
-        DisplayName="RichTextControls.Tests"
+        DisplayName="RichTextControls.ReadersExample"
         Square150x150Logo="Assets\Square150x150Logo.png"
         Square44x44Logo="Assets\Square44x44Logo.png"
-        Description="RichTextControls.Tests"
+        Description="RichTextControls.ReadersExample"
         BackgroundColor="transparent">
         <uap:DefaultTile Wide310x150Logo="Assets\Wide310x150Logo.png"/>
         <uap:SplashScreen Image="Assets\SplashScreen.png" />
       </uap:VisualElements>
     </Application>
   </Applications>
+
   <Capabilities>
-    <Capability Name="internetClientServer" />
+    <Capability Name="internetClient" />
   </Capabilities>
 </Package>
