--- RichTextControls/RichTextControls.Tests/RichTextBlockHtmlTestPage.xaml
@@ -1,17 +1,0 @@
-﻿<Page
-    x:Class="RichTextControls.Tests.RichTextBlockHtmlTestPage"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:RichTextControls.Tests"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:richtextcontrols="using:RichTextControls"
-    mc:Ignorable="d"
-    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
-
-    <Grid>
-        <ScrollViewer>
-        <RichTextBlock x:Name="abcd" richtextcontrols:Properties.Html="{x:Bind Path=Html}" x:DefaultBindMode="TwoWay"></RichTextBlock>
-        </ScrollViewer>
-    </Grid>
-</Page>
