--- RichTextControls/RichTextControls.Tests/RichTextControlTest.xaml
@@ -1,66 +1,0 @@
-﻿<Page
-    x:Class="RichTextControls.Tests.RichTextControlTest"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:RichTextControls.Tests"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:richtextcontrols="using:RichTextControls"
-    mc:Ignorable="d"
-    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
-
-        <Grid x:Name="testGrid">
-            
-            <StackPanel>
-                <Viewbox StretchDirection="DownOnly">
-                    <StackPanel Orientation="Horizontal">
-                        <TextBox x:Name="UrlBox" Width="115" HorizontalAlignment="Left" Margin="20,20,0,20" VerticalAlignment="Stretch"></TextBox>
-                        <Button x:Name="ReaderSharpButton" Content="ReaderSharp" Margin="20" Click="ReaderSharpButton_Click"></Button>
-                        <Button x:Name="ShowHtmlButton" Content="显示转换后HTML">
-                            <Button.Flyout>
-                                <Flyout>
-                                    <ScrollViewer Height="500" Width="500">
-                                        <richtextcontrols:CodeHighlightedTextBlock x:Name="CodeBlock" Code="{x:Bind Html}" Language="XML">
-                                        </richtextcontrols:CodeHighlightedTextBlock>
-                                    </ScrollViewer>
-                                </Flyout>
-                            </Button.Flyout>
-                        </Button>
-                        <Viewbox >
-                            <StackPanel>
-                                <StackPanel Orientation="Horizontal" VerticalAlignment="Center">
-                                    <TextBlock Text="宽度：" Margin="10"></TextBlock>
-                                    <Slider x:Name="ReaderWidthSlider" Width="200" VerticalAlignment="Center" ValueChanged="Slider_ValueChanged"/>
-                                </StackPanel>
-                                <StackPanel Orientation="Horizontal" VerticalAlignment="Center">
-                                    <TextBlock Text="字体大小：" Margin="10"></TextBlock>
-                                    <Slider x:Name="FontSizeSlider" Width="200" VerticalAlignment="Center" ValueChanged="FontSizeSlider_ValueChanged"/>
-                                </StackPanel>
-                            </StackPanel>
-                        </Viewbox>
-                        <DropDownButton x:Name="FontButton" Content="字体" Margin="30, 0, 0, 0">
-                            <DropDownButton.Flyout>
-                                <MenuFlyout x:Name="FontFlyout" >
-                                </MenuFlyout>
-                            </DropDownButton.Flyout>
-                        </DropDownButton>
-                    </StackPanel>
-                </Viewbox>
-                <Grid>
-                    <Grid.ColumnDefinitions>
-                        <ColumnDefinition Width="*"></ColumnDefinition>
-                        <ColumnDefinition Width="*"></ColumnDefinition>
-                    </Grid.ColumnDefinitions>
-                    <Viewbox  x:Name="TestGrid" Width="500" Height="500" Grid.Column="0" VerticalAlignment="Top">
-                        <!--<richtextcontrols:HtmlTextBlock x:Name="htmlTest" Html="{x:Bind ReaderHtml}" Opacity="0"></richtextcontrols:HtmlTextBlock>-->
-                        <RichTextBlock x:Name="htmlTest" Width="500" Height="500" richtextcontrols:Properties.Html="{x:Bind Html}" x:DefaultBindMode="TwoWay" SelectionChanged="HtmlTest_SelectionChanged" OverflowContentTarget="{Binding ElementName=fuckoff}" VerticalAlignment="Top"></RichTextBlock>
-                    </Viewbox>
-                    <RichTextBlockOverflow Grid.Column="1" x:Name="fuckoff"></RichTextBlockOverflow>
-                </Grid>
-
-
-            </StackPanel>
-        </Grid>
-
-
-</Page>
