+++ RichTextControls/RichTextControls.sln
@@ -1,15 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26430.6
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29102.190
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RichTextControls", "RichTextControls\RichTextControls.csproj", "{E3BD6F6F-B39E-4180-8182-DFE85C780BF4}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RichTextControls.ExampleApp", "RichTextControls.ExampleApp\RichTextControls.ExampleApp.csproj", "{3808E366-33EE-4E33-A572-02420DBDA1C9}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RichTextControls.Tests", "RichTextControls.Tests\RichTextControls.Tests.csproj", "{39C8ACC2-57D8-483D-9840-CA917FD28AB8}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RichTextControls.RichTextBlockConverter", "RichTextControls.RichTextBlockConverter\RichTextControls.RichTextBlockConverter.csproj", "{268C54E2-92E8-4014-8D5E-E22124CF31DC}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RichTextControls.ReadersExample", "RichTextControls.ReadersExample\RichTextControls.ReadersExample.csproj", "{1A771320-F255-4814-A363-E032A994874F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -66,46 +64,32 @@ Global
 		{3808E366-33EE-4E33-A572-02420DBDA1C9}.Release|x86.ActiveCfg = Release|x86
 		{3808E366-33EE-4E33-A572-02420DBDA1C9}.Release|x86.Build.0 = Release|x86
 		{3808E366-33EE-4E33-A572-02420DBDA1C9}.Release|x86.Deploy.0 = Release|x86
-		{39C8ACC2-57D8-483D-9840-CA917FD28AB8}.Debug|Any CPU.ActiveCfg = Debug|x86
-		{39C8ACC2-57D8-483D-9840-CA917FD28AB8}.Debug|Any CPU.Build.0 = Debug|x86
-		{39C8ACC2-57D8-483D-9840-CA917FD28AB8}.Debug|ARM.ActiveCfg = Debug|ARM
-		{39C8ACC2-57D8-483D-9840-CA917FD28AB8}.Debug|ARM.Build.0 = Debug|ARM
-		{39C8ACC2-57D8-483D-9840-CA917FD28AB8}.Debug|ARM.Deploy.0 = Debug|ARM
-		{39C8ACC2-57D8-483D-9840-CA917FD28AB8}.Debug|ARM64.ActiveCfg = Debug|x86
-		{39C8ACC2-57D8-483D-9840-CA917FD28AB8}.Debug|x64.ActiveCfg = Debug|x64
-		{39C8ACC2-57D8-483D-9840-CA917FD28AB8}.Debug|x64.Build.0 = Debug|x64
-		{39C8ACC2-57D8-483D-9840-CA917FD28AB8}.Debug|x64.Deploy.0 = Debug|x64
-		{39C8ACC2-57D8-483D-9840-CA917FD28AB8}.Debug|x86.ActiveCfg = Debug|x86
-		{39C8ACC2-57D8-483D-9840-CA917FD28AB8}.Debug|x86.Build.0 = Debug|x86
-		{39C8ACC2-57D8-483D-9840-CA917FD28AB8}.Debug|x86.Deploy.0 = Debug|x86
-		{39C8ACC2-57D8-483D-9840-CA917FD28AB8}.Release|Any CPU.ActiveCfg = Release|x86
-		{39C8ACC2-57D8-483D-9840-CA917FD28AB8}.Release|ARM.ActiveCfg = Release|ARM
-		{39C8ACC2-57D8-483D-9840-CA917FD28AB8}.Release|ARM.Build.0 = Release|ARM
-		{39C8ACC2-57D8-483D-9840-CA917FD28AB8}.Release|ARM.Deploy.0 = Release|ARM
-		{39C8ACC2-57D8-483D-9840-CA917FD28AB8}.Release|ARM64.ActiveCfg = Release|x86
-		{39C8ACC2-57D8-483D-9840-CA917FD28AB8}.Release|x64.ActiveCfg = Release|x64
-		{39C8ACC2-57D8-483D-9840-CA917FD28AB8}.Release|x64.Deploy.0 = Release|x64
-		{39C8ACC2-57D8-483D-9840-CA917FD28AB8}.Release|x86.ActiveCfg = Release|x86
-		{39C8ACC2-57D8-483D-9840-CA917FD28AB8}.Release|x86.Build.0 = Release|x86
-		{39C8ACC2-57D8-483D-9840-CA917FD28AB8}.Release|x86.Deploy.0 = Release|x86
-		{268C54E2-92E8-4014-8D5E-E22124CF31DC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{268C54E2-92E8-4014-8D5E-E22124CF31DC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{268C54E2-92E8-4014-8D5E-E22124CF31DC}.Debug|ARM.ActiveCfg = Debug|ARM
-		{268C54E2-92E8-4014-8D5E-E22124CF31DC}.Debug|ARM.Build.0 = Debug|ARM
-		{268C54E2-92E8-4014-8D5E-E22124CF31DC}.Debug|ARM64.ActiveCfg = Debug|ARM64
-		{268C54E2-92E8-4014-8D5E-E22124CF31DC}.Debug|ARM64.Build.0 = Debug|ARM64
-		{268C54E2-92E8-4014-8D5E-E22124CF31DC}.Debug|x64.ActiveCfg = Debug|x64
-		{268C54E2-92E8-4014-8D5E-E22124CF31DC}.Debug|x64.Build.0 = Debug|x64
-		{268C54E2-92E8-4014-8D5E-E22124CF31DC}.Debug|x86.ActiveCfg = Debug|x86
-		{268C54E2-92E8-4014-8D5E-E22124CF31DC}.Debug|x86.Build.0 = Debug|x86
-		{268C54E2-92E8-4014-8D5E-E22124CF31DC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{268C54E2-92E8-4014-8D5E-E22124CF31DC}.Release|ARM.ActiveCfg = Release|ARM
-		{268C54E2-92E8-4014-8D5E-E22124CF31DC}.Release|ARM.Build.0 = Release|ARM
-		{268C54E2-92E8-4014-8D5E-E22124CF31DC}.Release|ARM64.ActiveCfg = Release|ARM64
-		{268C54E2-92E8-4014-8D5E-E22124CF31DC}.Release|ARM64.Build.0 = Release|ARM64
-		{268C54E2-92E8-4014-8D5E-E22124CF31DC}.Release|x64.ActiveCfg = Release|x64
-		{268C54E2-92E8-4014-8D5E-E22124CF31DC}.Release|x86.ActiveCfg = Release|x86
-		{268C54E2-92E8-4014-8D5E-E22124CF31DC}.Release|x86.Build.0 = Release|x86
+		{1A771320-F255-4814-A363-E032A994874F}.Debug|Any CPU.ActiveCfg = Debug|x86
+		{1A771320-F255-4814-A363-E032A994874F}.Debug|ARM.ActiveCfg = Debug|ARM
+		{1A771320-F255-4814-A363-E032A994874F}.Debug|ARM.Build.0 = Debug|ARM
+		{1A771320-F255-4814-A363-E032A994874F}.Debug|ARM.Deploy.0 = Debug|ARM
+		{1A771320-F255-4814-A363-E032A994874F}.Debug|ARM64.ActiveCfg = Debug|ARM64
+		{1A771320-F255-4814-A363-E032A994874F}.Debug|ARM64.Build.0 = Debug|ARM64
+		{1A771320-F255-4814-A363-E032A994874F}.Debug|ARM64.Deploy.0 = Debug|ARM64
+		{1A771320-F255-4814-A363-E032A994874F}.Debug|x64.ActiveCfg = Debug|x64
+		{1A771320-F255-4814-A363-E032A994874F}.Debug|x64.Build.0 = Debug|x64
+		{1A771320-F255-4814-A363-E032A994874F}.Debug|x64.Deploy.0 = Debug|x64
+		{1A771320-F255-4814-A363-E032A994874F}.Debug|x86.ActiveCfg = Debug|x86
+		{1A771320-F255-4814-A363-E032A994874F}.Debug|x86.Build.0 = Debug|x86
+		{1A771320-F255-4814-A363-E032A994874F}.Debug|x86.Deploy.0 = Debug|x86
+		{1A771320-F255-4814-A363-E032A994874F}.Release|Any CPU.ActiveCfg = Release|x86
+		{1A771320-F255-4814-A363-E032A994874F}.Release|ARM.ActiveCfg = Release|ARM
+		{1A771320-F255-4814-A363-E032A994874F}.Release|ARM.Build.0 = Release|ARM
+		{1A771320-F255-4814-A363-E032A994874F}.Release|ARM.Deploy.0 = Release|ARM
+		{1A771320-F255-4814-A363-E032A994874F}.Release|ARM64.ActiveCfg = Release|ARM64
+		{1A771320-F255-4814-A363-E032A994874F}.Release|ARM64.Build.0 = Release|ARM64
+		{1A771320-F255-4814-A363-E032A994874F}.Release|ARM64.Deploy.0 = Release|ARM64
+		{1A771320-F255-4814-A363-E032A994874F}.Release|x64.ActiveCfg = Release|x64
+		{1A771320-F255-4814-A363-E032A994874F}.Release|x64.Build.0 = Release|x64
+		{1A771320-F255-4814-A363-E032A994874F}.Release|x64.Deploy.0 = Release|x64
+		{1A771320-F255-4814-A363-E032A994874F}.Release|x86.ActiveCfg = Release|x86
+		{1A771320-F255-4814-A363-E032A994874F}.Release|x86.Build.0 = Release|x86
+		{1A771320-F255-4814-A363-E032A994874F}.Release|x86.Deploy.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
